forced data to consist of doubles
svn path=/trunk/matlab/; revision=2970


Former-commit-id: 9e4be41a12df0cd7e0de8e92d3daad8fe0ffb696

+++ applications/Rijkswaterstaat/jarkus/jarkus_transects.m
@@ -224,7 +224,7 @@ for i = datasetOrder'
             end
             
             % retrieve data
-            data = nc_varget(OPT.url, name, start-1, count);
+            data = double(nc_varget(OPT.url, name, start-1, count));
             
             % set precision
             if ~ischar(data)
